+++ src/main/java/ssu/eatssu/domain/repository/MealMenuRepository.java
@@ -1,7 +1,7 @@
 package ssu.eatssu.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ssu.eatssu.domain.MealMenu;
+import ssu.eatssu.domain.menu.MealMenu;
 
 public interface MealMenuRepository extends JpaRepository<MealMenu, Long> {
 }
