+++ src/main/java/ssu/eatssu/domain/repository/MealRepository.java
@@ -1,8 +1,8 @@
 package ssu.eatssu.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ssu.eatssu.domain.Meal;
-import ssu.eatssu.domain.Restaurant;
+import ssu.eatssu.domain.menu.Meal;
+import ssu.eatssu.domain.restaurant.Restaurant;
 import ssu.eatssu.domain.enums.TimePart;
 
 import java.util.Date;
