+++ src/main/java/ssu/eatssu/domain/repository/MenuRepository.java
@@ -1,8 +1,8 @@
 package ssu.eatssu.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ssu.eatssu.domain.Menu;
-import ssu.eatssu.domain.Restaurant;
+import ssu.eatssu.domain.menu.Menu;
+import ssu.eatssu.domain.restaurant.Restaurant;
 
 import java.util.List;
 import java.util.Optional;
