+++ src/main/java/ssu/eatssu/domain/repository/ReviewImageRepository.java
@@ -1,7 +1,7 @@
 package ssu.eatssu.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ssu.eatssu.domain.ReviewImage;
+import ssu.eatssu.domain.review.ReviewImage;
 
 public interface ReviewImageRepository extends JpaRepository<ReviewImage, Long> {
 }
