+++ src/main/java/ssu/eatssu/domain/repository/ReviewReportRepository.java
@@ -1,7 +1,7 @@
 package ssu.eatssu.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ssu.eatssu.domain.ReviewReport;
+import ssu.eatssu.domain.review.ReviewReport;
 
 public interface ReviewReportRepository extends JpaRepository<ReviewReport, Long>{
 }
