+++ src/main/java/ssu/eatssu/domain/repository/ReviewRepository.java
@@ -1,7 +1,7 @@
 package ssu.eatssu.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ssu.eatssu.domain.Menu;
+import ssu.eatssu.domain.menu.Menu;
 import ssu.eatssu.domain.review.Review;
 
 import java.util.List;
