+++ src/main/java/ssu/eatssu/domain/repository/ReviewRepositoryImpl.java
@@ -12,7 +12,10 @@ import ssu.eatssu.domain.*;
 
 import java.util.ArrayList;
 import java.util.List;
+import ssu.eatssu.domain.menu.Meal;
+import ssu.eatssu.domain.menu.Menu;
 import ssu.eatssu.domain.review.Review;
+import ssu.eatssu.domain.user.User;
 
 @Repository
 @RequiredArgsConstructor
