+++ src/main/java/ssu/eatssu/domain/repository/UserInquiryRepository.java
@@ -1,7 +1,7 @@
 package ssu.eatssu.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ssu.eatssu.domain.UserInquiry;
+import ssu.eatssu.domain.user.UserInquiry;
 
 public interface UserInquiryRepository extends JpaRepository<UserInquiry, Long>{
 }
