+++ src/main/java/ssu/eatssu/domain/repository/UserRepository.java
@@ -1,7 +1,7 @@
 package ssu.eatssu.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ssu.eatssu.domain.User;
+import ssu.eatssu.domain.user.User;
 
 import java.util.Optional;
