+++ src/main/java/ssu/eatssu/domain/review/ReviewImage.java
@@ -1,4 +1,4 @@
-package ssu.eatssu.domain;
+package ssu.eatssu.domain.review;
 
 import jakarta.persistence.*;
 import lombok.*;
