+++ src/main/java/ssu/eatssu/domain/user/OauthProvider.java
@@ -1,4 +1,4 @@
-package ssu.eatssu.domain.enums;
+package ssu.eatssu.domain.user;
 
 public enum OauthProvider {
     EATSSU, KAKAO, APPLE;
