+++ src/main/java/ssu/eatssu/domain/user/Role.java
@@ -1,4 +1,4 @@
-package ssu.eatssu.domain.enums;
+package ssu.eatssu.domain.user;
 
 import lombok.Getter;
 import org.springframework.security.core.GrantedAuthority;
