+++ src/main/java/ssu/eatssu/domain/user/UserStatus.java
@@ -1,4 +1,4 @@
-package ssu.eatssu.domain.enums;
+package ssu.eatssu.domain.user;
 
 import lombok.Getter;
