+++ src/main/java/ssu/eatssu/service/MenuService.java
@@ -4,11 +4,11 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import ssu.eatssu.domain.Meal;
-import ssu.eatssu.domain.MealMenu;
-import ssu.eatssu.domain.Menu;
-import ssu.eatssu.domain.Restaurant;
-import ssu.eatssu.domain.enums.RestaurantName;
+import ssu.eatssu.domain.menu.Meal;
+import ssu.eatssu.domain.menu.MealMenu;
+import ssu.eatssu.domain.menu.Menu;
+import ssu.eatssu.domain.restaurant.Restaurant;
+import ssu.eatssu.domain.restaurant.RestaurantName;
 import ssu.eatssu.domain.enums.TimePart;
 import ssu.eatssu.domain.repository.MealMenuRepository;
 import ssu.eatssu.domain.repository.MealRepository;
