+++ src/main/java/ssu/eatssu/service/OauthService.java
@@ -10,18 +10,16 @@ import io.jsonwebtoken.Jwts;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
-import ssu.eatssu.domain.User;
-import ssu.eatssu.domain.enums.OauthProvider;
+import ssu.eatssu.domain.user.User;
+import ssu.eatssu.domain.user.OauthProvider;
 import ssu.eatssu.domain.repository.UserRepository;
 import ssu.eatssu.handler.response.BaseException;
 import ssu.eatssu.handler.response.BaseResponseStatus;
-import ssu.eatssu.jwt.JwtTokenProvider;
 import ssu.eatssu.service.vo.OauthInfo;
 import ssu.eatssu.web.oauth.dto.AppleKeys;
 import ssu.eatssu.web.user.dto.Tokens;
