+++ src/main/java/ssu/eatssu/service/ReportService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ssu.eatssu.domain.review.Review;
-import ssu.eatssu.domain.ReviewReport;
-import ssu.eatssu.domain.User;
+import ssu.eatssu.domain.review.ReviewReport;
+import ssu.eatssu.domain.user.User;
 import ssu.eatssu.domain.enums.ReportStatus;
 import ssu.eatssu.domain.repository.ReviewReportRepository;
 import ssu.eatssu.domain.repository.ReviewRepository;
