+++ src/main/java/ssu/eatssu/service/ReviewService.java
@@ -7,9 +7,12 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-import ssu.eatssu.domain.*;
+import ssu.eatssu.domain.menu.Meal;
+import ssu.eatssu.domain.menu.Menu;
 import ssu.eatssu.domain.repository.*;
 import ssu.eatssu.domain.review.Review;
+import ssu.eatssu.domain.review.ReviewImage;
+import ssu.eatssu.domain.user.User;
 import ssu.eatssu.handler.response.BaseException;
 import ssu.eatssu.utils.S3Uploader;
 import ssu.eatssu.utils.SecurityUtil;
