+++ src/main/java/ssu/eatssu/service/UserInquiryService.java
@@ -3,8 +3,8 @@ package ssu.eatssu.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ssu.eatssu.domain.User;
-import ssu.eatssu.domain.UserInquiry;
+import ssu.eatssu.domain.user.User;
+import ssu.eatssu.domain.user.UserInquiry;
 import ssu.eatssu.domain.repository.UserInquiryRepository;
 import ssu.eatssu.domain.repository.UserRepository;
 import ssu.eatssu.handler.response.BaseException;
