+++ src/main/java/ssu/eatssu/service/UserService.java
@@ -9,9 +9,9 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import ssu.eatssu.domain.review.Review;
-import ssu.eatssu.domain.ReviewReport;
-import ssu.eatssu.domain.User;
-import ssu.eatssu.domain.enums.OauthProvider;
+import ssu.eatssu.domain.review.ReviewReport;
+import ssu.eatssu.domain.user.User;
+import ssu.eatssu.domain.user.OauthProvider;
 import ssu.eatssu.domain.repository.ReviewReportRepository;
 import ssu.eatssu.domain.repository.UserRepository;
 import ssu.eatssu.handler.response.BaseResponseStatus;
