+++ src/main/java/ssu/eatssu/slack/SlackMessageFormat.java
@@ -2,9 +2,9 @@ package ssu.eatssu.slack;
 
 import org.springframework.stereotype.Component;
 import ssu.eatssu.domain.review.Review;
-import ssu.eatssu.domain.ReviewReport;
-import ssu.eatssu.domain.User;
-import ssu.eatssu.domain.UserInquiry;
+import ssu.eatssu.domain.review.ReviewReport;
+import ssu.eatssu.domain.user.User;
+import ssu.eatssu.domain.user.UserInquiry;
 
 import java.text.MessageFormat;
