+++ src/main/java/ssu/eatssu/web/mypage/dto/MypageInfo.java
@@ -4,7 +4,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import ssu.eatssu.domain.enums.OauthProvider;
+import ssu.eatssu.domain.user.OauthProvider;
 
 @AllArgsConstructor
 @Builder
