+++ src/main/java/ssu/eatssu/web/report/ReportController.java
@@ -9,11 +9,10 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
-import ssu.eatssu.domain.ReviewReport;
+import ssu.eatssu.domain.review.ReviewReport;
 import ssu.eatssu.domain.enums.ReviewReportType;
 import ssu.eatssu.domain.repository.ReviewRepository;
 import ssu.eatssu.domain.repository.UserRepository;
-import ssu.eatssu.handler.response.BaseException;
 import ssu.eatssu.handler.response.BaseResponse;
 import ssu.eatssu.service.ReportService;
 import ssu.eatssu.slack.SlackChannel;
